+++ script/go_build.sh
@@ -7,9 +7,8 @@
 [ -z "$GITBRANCH" ] && GITBRANCH=$(git rev-parse --abbrev-ref HEAD 2>/dev/null)
 
 go_build() {
-  [ -d "${DIST}" ] && rm -rf "${DIST}/*"
+  [ -d "${DIST}" ] && rm -rf "${DIST:?}/*"
   [ -d "${DIST}" ] || mkdir -p "${DIST}"
-  glide install -v
   CGO_ENABLED=0 go build \
     -ldflags "-X main.Version=${VERSION} -X main.GitCommit=${GITCOMMIT} -X main.GitBranch=${GITBRANCH} -X main.BuildTime=${BUILDTIME}" \
     -v -o "${DIST}/pumba"
