+++ script/gox_build.sh
@@ -10,10 +10,9 @@ exarch="amd64 386"
 oslist="linux windows darwin"
 
 gox_build() {
-  [ -d "${DIST}" ] && rm -rf "${DIST}/*"
+  [ -d "${DIST}" ] && rm -rf "${DIST:?}/*"
   [ -d "${DIST}" ] || mkdir -p "${DIST}"
-  echo "Building" ${BUILD_VERSION} "on" ${BUILD_DATE}
-  glide install -v
+  echo "Building ${BUILD_VERSION} on ${BUILD_DATE}"
   gox -os="${oslist}" -arch="${exarch}" -cgo=false \
     -ldflags "-X main.Version=${VERSION} -X main.GitCommit=${GITCOMMIT} -X main.GitBranch=${GITBRANCH} -X main.BuildTime=${BUILDTIME}" \
     -verbose -output="${DIST}/pumba_{{.OS}}_{{.Arch}}" .
